Refactor
Following modules were created
1. Annotation
2. Tiling
3. Validation
4. Discovery

Scripts have also been modified to use the updated locations.

Inference and training remain in the main directory because they are run
within a docker.
