+++ detectron2/discovery/__init__.py
@@ -1,0 +1,1 @@
+from .discovery import DataDiscovery
