+++ detectron2/discovery/discovery.py
