--- detectron2/tiles_on_tiff.py
@@ -1,1 +1,0 @@
-
