+++ detectron2/tiling/__init__.py
@@ -1,0 +1,1 @@
+from .tile_generator import TileGenerator
