+++ detectron2/validation/geometry.py
